requirements.txt
Flask==2.2.5
firebase-admin==6.0.1
bcrypt==4.0.1
PyJWT==2.7.0
flask-limiter
